# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/user/GetUser":{"post":{"operationId":"UserController_nothing","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"properties":{"username":{"type":"string"}}}}}},"responses":{"201":{"description":""}}}},"/user/GetAllConfigs":{"post":{"operationId":"UserController_getconfigs","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"properties":{"username":{"type":"string"}}}}}},"responses":{"201":{"description":""}}}},"/user/GetConfig":{"post":{"operationId":"UserController_getconfig","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"properties":{"username":{"type":"string"},"name":{"type":"string"}}}}}},"responses":{"201":{"description":""}}}},"/user/CreateConfig":{"post":{"operationId":"UserController_createconfig","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"properties":{"username":{"type":"string"},"name":{"type":"string"},"content":{"type":"string"}}}}}},"responses":{"201":{"description":""}}}},"/user/DeleteConfig":{"post":{"operationId":"UserController_deleteconfig","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"properties":{"username":{"type":"string"},"name":{"type":"string"}}}}}},"responses":{"201":{"description":""}}}},"/user/UpdateConfig":{"post":{"operationId":"UserController_updateconfig","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"properties":{"username":{"type":"string"},"name":{"type":"string"},"content":{"type":"string"}}}}}},"responses":{"201":{"description":""}}}},"/user/GetAllLuas":{"post":{"operationId":"UserController_getlaus","parameters":[],"responses":{"201":{"description":""}}}}},"info":{"title":"Cats example","description":"The cats API description","version":"1.0","contact":{}},"tags":[{"name":"cats","description":""}],"servers":[],"components":{"schemas":{}}}